requirements.txt
asttokens==2.4.1
attrs==23.2.0
aws-cdk-lib==2.144.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
cattrs==23.2.3
colorama==0.4.6
comm==0.2.2
constructs==10.3.0
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jsii==1.99.0
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess==0.7.0
publication==0.0.3
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typeguard==2.13.3
typing_extensions==4.12.1
tzdata==2024.1
wcwidth==0.2.13
